week-5/monitoring/evidently/requirements.txt
evidently
clickhouse-driver
pandas
python-dotenv
requests 